Handle inflection across short durations.
Fixes issue where the inflection between very short phrases could
be skipped.

+++ tests/choreograph/Configuration.h
@@ -38,6 +38,6 @@ namespace choreograph
 /// Switch to double if you need more precision.
 ///
 
-using Time = float;
+using Time = double;
 
 } // namespace choreograph
